added stub mode for development

+++ src/mobx-stores/Assets.store.js
@@ -1,10 +1,13 @@
 import { observable } from 'mobx'
-import { axios, generateLine, toFormData } from 'common'
-import { create as hydrate, persist } from 'mobx-persist'
+import { axios, generateDemoTable, generateLine, toFormData } from 'common'
+import { persist } from 'mobx-persist'
+import labels from './labels.list'
 
-// const delay = () => new Promise(resolve => setTimeout(resolve, 1000))
+const PAGE_SIZE = 10
+const delay = () => new Promise(resolve => setTimeout(resolve, 1000))
 
-class AssetsStore {
+export default class AssetsStore {
+  labels = labels
   @persist('list') @observable list = []
   @persist @observable totalPages = 1
   @persist @observable totalElements = 1
@@ -13,6 +16,7 @@ class AssetsStore {
   @persist('object') @observable paging = { page: 0 }
 
   @persist @observable activeId = null
+  @persist @observable stub = false
   @persist('object') @observable active = {}
   @persist('object') @observable activeItem = {}
 
@@ -46,49 +50,6 @@ class AssetsStore {
       return map
     }), {})
   }
-
-  /**
-   Current java mapping for this fields:
-   @Length(max = 50) @NotBlank private String name;
-   @Length(max = 50) @NotBlank private String type;
-   @Length(max = 50) private String barcode;
-   @Length(max = 50) private String number;
-   @Length(max = 50) private String serial;
-   @Length(max = 1000) private String note;
-   @Length(max = 1000) private String description;
-   @Length(max = 50) private String department;
-   @Length(max = 50) private String manufacture;
-   @Length(max = 50) private String model;
-   @Length(max = 50) private String rfid;
-   private LocalDateTime lastUsedDate;
-   @Valid private KeyLocationDto keyLocation;
-   private ImageDto image;
-   */
-
-  labels = [
-    { label: 'id', key: 'id', hidden: true },
-    { label: 'Asset Name', key: 'name', required: true, hideOnView: true },
-    { label: 'Asset Number', key: 'number', required: true },
-    { label: 'Asset Type', key: 'type', required: true, defaultVisible: true, searchOrder: 1 },
-    { label: 'Owner/Department', key: 'department', defaultVisible: true, searchOrder: 5 },
-    { label: 'Location', key: 'keyLocation', defaultVisible: true, searchOrder: 2 },
-    { label: 'Model', key: 'model', searchOrder: 3 },
-    { label: 'Manufacturer', key: 'manufacturer', searchOrder: 4 },
-    { label: 'Description', key: 'description' },
-    { label: 'Search Terms', key: 'search_terms' },
-    { label: 'RFID Assigned', key: 'rfidAssigned', searchOrder: 6 },
-    { label: 'Serial Number', key: 'serial' },
-    { label: 'Barcode Number', key: 'barcode', required: true },
-    { label: 'RFID Number', key: 'rfid' },
-    { label: 'Notes', key: 'note', multiline: true },
-    { label: 'Image', key: 'image', hidden: true },
-    {
-      label: 'Update Location Date', key: 'lastUsedDate', hideOnEdit: true, searchOrder: 7, dateFilterKeys: [
-      { key: 'fromUpdateLocationDate', label: 'Last Update Date from' },
-      { key: 'toUpdateLocationDate', label: 'Last Update Date to' }
-    ]
-    }
-  ]
 
   getXlsxLabels() {
     if ( this.xlsTable.length && this.xlsTable[0] ) {
@@ -125,9 +86,14 @@ class AssetsStore {
 
   async remove(ids) {
     ids = Array.isArray(ids) ? ids : [ids]
-    await Promise.all(ids.map(id => {
-      return axios.delete(`/api/v1/hospital/assets/${ id }`)
-    }))
+    if ( !this.stub ) {
+      await Promise.all(ids.map(id => {
+        return axios.delete(`/api/v1/hospital/assets/${ id }`)
+      }))
+    } else {
+      await delay()
+    }
+
     this.list = this.list.filter(({ id }) => !ids.includes(id))
   }
 
@@ -137,11 +103,19 @@ class AssetsStore {
     delete assetData.keyLocation
     delete assetData.lastUsedDate
 
-    const { data } = await axios.post('/api/v1/hospital/assets', toFormData(assetData))
-    this.list.push(data)
-    this.active = data
-    this.activeItem = data
-    return data
+    let newItem
+    if ( this.stub ) {
+      await delay()
+      newItem = { ...this.active }
+    } else {
+      const { data } = await axios.post('/api/v1/hospital/assets', toFormData(assetData))
+      newItem = data
+    }
+
+    this.list.push(newItem)
+    this.active = newItem
+    this.activeItem = newItem
+    return newItem
   }
 
   setRandomForActive() {
@@ -155,8 +129,16 @@ class AssetsStore {
     delete assetData.keyLocation
     delete assetData.lastUsedDate
     // delete assetData.id
-    const { data } = await axios.put(`/api/v1/hospital/assets/${ assetData.id }`, toFormData(assetData))
-    return Object.assign(this.activeItem, this.active, data)
+    let updatedItem
+    if ( this.stub ) {
+      await delay()
+      updatedItem = assetData
+    } else {
+      const { data } = await axios.put(`/api/v1/hospital/assets/${ assetData.id }`, toFormData(assetData))
+      updatedItem = data
+    }
+
+    return Object.assign(this.activeItem, this.active, updatedItem)
   }
 
   async changeSort(key) {
@@ -170,11 +152,14 @@ class AssetsStore {
 
   async loadAutocompleteValues(key = 'department', query) {
     if ( key === 'rfidAssigned' ) {
-      return await Promise.resolve([
+      return await [
         'All Assets',
         'RFID Assigned',
         'RFID Not Assigned'
-      ])
+      ]
+    } else if ( this.stub ) {
+      await delay()
+      return []
     } else {
       let params = null
       if ( query ) {
@@ -188,25 +173,21 @@ class AssetsStore {
   async loadList() {
     this.tableLoading = true
 
-    let sort
-    if ( this.sort.key ) {
-      sort = `${ this.sort.key },${ this.sort.asc ? 'asc' : 'desc' }`
-    }
-    const params = { ...this.searchParams, ...this.paging, size: 10, sort }
-    const { data: { content, totalPages, totalElements } } = await axios.get('/api/v1/hospital/assets', { params })
+    if ( this.stub ) {
+      const content = generateDemoTable(this.labels, PAGE_SIZE)
+      await delay()
+      Object.assign(this, { list: content, totalPages: 1, totalElements: content.length })
+    } else {
+      let sort
+      if ( this.sort.key ) {
+        sort = `${ this.sort.key },${ this.sort.asc ? 'asc' : 'desc' }`
+      }
+      const params = { ...this.searchParams, ...this.paging, size: PAGE_SIZE, sort }
 
-    this.list = content
-    this.totalPages = totalPages
-    this.totalElements = totalElements
+      const { data: { content, totalPages, totalElements } } = await axios.get('/api/v1/hospital/assets', { params })
+      Object.assign(this, { list: content, totalPages, totalElements })
+    }
+
     this.tableLoading = false
-    this.loadAuto()
   }
 }
-
-const assetsStore = new AssetsStore()
-
-hydrate()('assetsStore', assetsStore).then(() => {
-  console.log('assetsStore hydrated')
-})
-
-export default assetsStore
